svn merge -r 35038:HEAD http://code.sci.utah.edu/cibc/src/ Note: To get painter running, I had to comment out calls to SCIRun::clip_nodes and SCIRun::clip_faces on line 250 and 258 of Core/Events/OpenGLViewer.cc because their code resides in the as-yet un-compiled Core/Algorithms/Fields directoriy.

+++ src/Core/Events/EventManager.cc
@@ -279,7 +279,8 @@ EventManager::run()
     if (event->get_time()) {
       last_event_time = event->get_time();
     } else if (last_event_time) {
-      last_event_time += (timer.time() - last_timer_time)*100.0;
+      last_event_time +=
+        (unsigned long)((timer.time() - last_timer_time) * 100.0);
       event->set_time(last_event_time);
     }
     last_timer_time = timer.time();
